% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ess.R
\name{calc_ess}
\alias{calc_ess}
\title{Calculate effective sample size}
\usage{
calc_ess(newgamma, oldgamma, part_w, part_loglike)
}
\arguments{
\item{newgamma}{new temperature}

\item{oldgamma}{old temperature}

\item{part_w}{particule weights}

\item{part_loglike}{particule log likelihood}
}
\value{
A list of arguments defining the problem.
}
\description{
Calculates the effective sample size
}
