% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_args_logistic_growth.R
\name{define_args_logistic_growth}
\alias{define_args_logistic_growth}
\title{Arguments for SMCfeatures}
\usage{
define_args_logistic_growth()
}
\value{
A list of arguments defining the problem.
}
\description{
Extract arguments necessary to run SMCfeatures
}
