% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_Gaussian.R
\name{log_likelihood_Gaussian}
\alias{log_likelihood_Gaussian}
\title{Log-Likelihood function for the logistic growth case study}
\usage{
log_likelihood_Gaussian(parameters, simulation, args)
}
\arguments{
\item{parameters}{vector of parameters, with last parameter the standard deviation}

\item{simulation}{list like object that contains output_data (output of simulation)}

\item{args}{a list like object of arguments that includes a boolean include_data_constraints, that indicates if data constraints should be included, and output_data (output of observed data)}
}
\value{
log likelihood of simulation outputs against empirical data if args$include_data_constraints set to TRUE, 0 otherwise.
}
\description{
Gaussian likelihood function
}
