% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_for_plots_logistic_growth.R
\name{model_for_plots_logistic_growth}
\alias{model_for_plots_logistic_growth}
\title{Model simulation on input data}
\usage{
model_for_plots_logistic_growth(parameters, args)
}
\arguments{
\item{parameters}{vector of parameters (growth rate, capacity, initial abundance)}

\item{args}{a list of arguments as defined in \link[SMCfeatures]{SMCfeatures}}
}
\value{
ist like object representing simulation outputs
t_data: time steps for data (args$input_data)
y_data: output of simulation on input data
}
\description{
Simulate outputs of calibrated model, on data points only
}
