% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logistic_growth_constraint.R
\name{model_logistic_growth_constraint}
\alias{model_logistic_growth_constraint}
\title{model_logistic_growth_constraint}
\usage{
model_logistic_growth_constraint(parameters)
}
\arguments{
\item{parameters}{vector of parameters (growth rate, capacity, initial abundance)}
}
\value{
simulation with added t_const, and y_const (output of simulation on input t_const)
}
\description{
Simulate
}
