% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logistic_growth_data.R
\name{model_logistic_growth_data}
\alias{model_logistic_growth_data}
\title{model_logistic_growth_data}
\usage{
model_logistic_growth_data(parameters, simulation, args)
}
\arguments{
\item{parameters}{vector of parameters (growth rate, capacity, initial abundance)}

\item{simulation}{list of simulation arguments}

\item{args}{a list of arguments as returned by \link[SMCfeatures]{args_function}}
}
\value{
simulation with added t_data, and y_data (output of simulation on input data)
}
\description{
Simulate
}
