% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_on_data.R
\name{model_on_data}
\alias{model_on_data}
\title{Model simulation on input data}
\usage{
model_on_data(parameters, simulation, args)
}
\arguments{
\item{parameters}{vector of parameters (growth rate, capacity, initial abundance)}

\item{simulation}{list of simulation arguments, as returned by \link[SMCfeatures]{model_logistic_growth_constraint}}

\item{args}{a list of arguments as defined in \link[SMCfeatures]{SMCfeatures}}
}
\value{
ist like object representing simulation outputs
t_data: time steps for data (args$input_data)
y_data: output of simulation on input data
}
\description{
Simulate outputs of calibrated model, on data points only
}
