% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_sampler.R
\name{uniform_sampler}
\alias{uniform_sampler}
\title{Uniform sampling of parameters}
\usage{
uniform_sampler(args)
}
\arguments{
\item{args}{a list of arguments as returned by \link[SMCfeatures]{define_args_logistic_growth}}
}
\value{
a vector of parameter samples
}
\description{
This function uniformly samples parameters between lower and upper bounds defined in args
}
