% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_transform.R
\name{uniform_transform}
\alias{uniform_transform}
\title{transformation function}
\usage{
uniform_transform(theta, args)
}
\arguments{
\item{theta}{vector of parameters to transform}

\item{args}{a list of arguments as returned by \link[SMCfeatures]{define_args_logistic_growth}}
}
\value{
vector of transformed parameters
}
\description{
This function uses a transform for uniform distributions, so all values are within the uniform bounds when sampling using SMC.
}
