% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_transform_inverse.R
\name{uniform_transform_inverse}
\alias{uniform_transform_inverse}
\title{inverse transformation function}
\usage{
uniform_transform_inverse(theta_trans, args)
}
\arguments{
\item{theta_trans}{vector of parameters to back-transform}

\item{args}{a list of arguments as returned by \link[SMCfeatures]{define_args_logistic_growth}}
}
\value{
vector of transformed parameters
}
\description{
This function uses a reverts the transformation for uniform distributions as defined in \link[SMCfeatures]{uniform_transform}
}
