% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrepancy_logistic_growth.R
\name{discrepancy_logistic_growth}
\alias{discrepancy_logistic_growth}
\title{Discrepancy function}
\usage{
discrepancy_logistic_growth(simulation, args)
}
\arguments{
\item{simulation}{list of simulation arguments}

\item{args}{a list of arguments as returned by \link[SMCfeatures]{args_function}}
}
\value{
summary statistic (discrepancy measure).
}
\description{
Discrepancy measure
}
