% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_logistic_growth.R
\name{log_likelihood_logistic_growth}
\alias{log_likelihood_logistic_growth}
\title{logLikelihood_logistic_growth}
\usage{
log_likelihood_logistic_growth(parameters, simulation, args)
}
\arguments{
\item{parameters}{parameters}

\item{simulation}{list of simulation arguments}

\item{args}{a list of arguments as returned by \link[SMCfeatures]{args_function}}
}
\value{
log likelihood
}
\description{
Gaussian likelihood function
}
