% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_combined.R
\name{MCMC_combined}
\alias{MCMC_combined}
\title{Monte Carlo Markov Chain}
\usage{
MCMC_combined(
  mcmc_steps,
  params_transformed,
  cov_matrix,
  args,
  disc,
  acc_count,
  dist_next,
  sims,
  loglike,
  newgamma
)
}
\arguments{
\item{cov_matrix}{covariance matrix}

\item{dist_next}{next objective distance}

\item{loglike}{log-likelihood}

\item{newgamma}{new temperature}

\item{i}{line number of particle to be moved}

\item{sim_args}{a list of arguments as returned by \link[EEMtoolbox]{args_function}}

\item{mcmc_trials}{number of MCMC steps to try before selecting appropriate number.}

\item{part_vals}{matrix of current values of parameters of each particle (each particle represented by a row)}

\item{part_s}{vector of discrepancy measures of each particle (each particle represented by a row)}

\item{part_sim}{matrix of current simulation values: equilibriums and eigenvalues of jacobian (each particle represented by a row)}

\item{summ_func}{function calculating equilibrium points and real parts of the Jacobians eigenvalues to summarise ecosystem features. Default =summarise_ecosystem_features_GLV. Options include summarise_ecosystem_features_Baker (automatically chosen if model="Bimler-Baker") and summarise_ecosystem_features_Gompertz, (automatically chosen if model="Gompertz"). Needs to be defined if model="customized" chosen.}

\item{disc_func}{summary statistic (discrepancy measure).}

\item{trans_finv}{inverse of trans_f function.}

\item{pdf}{joint probability density function.}
}
\value{
list
part_vals: updated value of parameter values for particle i
part_s: discrepancy measure for particle i
part_sim: summary of ecosystem features for particle i
i_acc: number of times particle movement accepted
sims_mcmc: number of successful walks
}
\description{
Runs MCMC
}
