% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_combined.R
\name{MCMC_combined}
\alias{MCMC_combined}
\title{Monte Carlo Markov Chain}
\usage{
MCMC_combined(
  args,
  mcmc_steps,
  params_transformed,
  cov_matrix,
  disc,
  dist_next,
  sims,
  loglike,
  acc_count,
  newgamma
)
}
\arguments{
\item{args}{a list of arguments as returned by \link[EEMtoolbox]{args_function}}

\item{mcmc_steps}{number of MCMC steps}

\item{params_transformed}{vector of values of parameters for current particle}

\item{cov_matrix}{covariance matrix}

\item{disc}{current discrepancy of the particle}

\item{dist_next}{next objective distance}

\item{sims}{list like object summarizing simulation outcomes}

\item{loglike}{log-likelihood of parameters}

\item{acc_count}{number of times the particle has been moved}

\item{newgamma}{new temperature}
}
\value{
list
params_transformed: vector of moved parameters
acc_count: number of times particle has been moved
sims: simulation outputs on expert constraint
loglike log likelihood of moved parameters
}
\description{
Runs MCMC
}
