% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logistic_growth.R
\name{model_logistic_growth}
\alias{model_logistic_growth}
\title{Model of logistic growth case study}
\usage{
model_logistic_growth(parameters, input)
}
\arguments{
\item{parameters}{vector of parameters (growth rate, capacity, initial abundance)}

\item{input}{input data. here it is a vector of time steps of interest}
}
\value{
y_data: output of simulation on input data as a vector
}
\description{
Simulate outputs of logistic growth model
}
