% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logistic_growth_constraint.R
\name{model_logistic_growth_constraint}
\alias{model_logistic_growth_constraint}
\title{model_logistic_growth_constraint}
\usage{
model_logistic_growth_constraint(parameters)
}
\arguments{
\item{parameters}{vector of parameters (growth rate, capacity, initial abundance, sigma)}
}
\value{
list like object representing simulation outputs
t_const: vector of time constraints (time step 5 and 50)
y_const: output of simulation on input t_const
K: carrying capacity
}
\description{
Simulate
}
