% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logistic_growth_plots.R
\name{model_logistic_growth_plots}
\alias{model_logistic_growth_plots}
\title{model_logistic_growth_plot}
\usage{
model_logistic_growth_plots(parameters, simulation, args)
}
\arguments{
\item{parameters}{vector of parameters (growth rate, capacity, initial abundance)}

\item{simulation}{list of simulation arguments}

\item{args}{a list of arguments as returned by \link[SMCfeatures]{args_function}}
}
\value{
simulation with added time, and y (output of simulation)
}
\description{
Simulate
}
