% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC_combined.R
\name{SMC_combined}
\alias{SMC_combined}
\title{Generation of model ensembles Sequential Monte Carlo - Approximate Bayesian Computation.}
\usage{
SMC_combined(
  args,
  n_particles,
  mcmc_trials,
  a_disc,
  a_like,
  c,
  p_acc_min,
  n_cores
)
}
\arguments{
\item{args}{a list of arguments as returned by \link[EEMtoolbox]{args_function}}
}
\value{
list: sims=number of simulations
part_vals=parameter values
part_s=discrepancy value
prior_sample=prior distribution
}
\description{
Generation of model ensembles using Adaptive sequential Monte Carlo for approximate Bayesian computation
}
