% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMCfeatures.R
\name{SMCfeatures}
\alias{SMCfeatures}
\title{SMCfeatures}
\usage{
SMCfeatures(
  upper,
  lower,
  simulate_model,
  calculate_discrepancy,
  simulate_data = SMCfeatures::model_data,
  input_data = NA,
  output_data = NA,
  calculate_log_likelihood = SMCfeatures::log_likelihood_Gaussian,
  include_expert_constraints = TRUE,
  include_data_constraints = TRUE,
  sampler = SMCfeatures::uniform_sampler,
  trans_f = SMCfeatures::uniform_transform,
  trans_finv = SMCfeatures::uniform_transform_inverse,
  pdf = SMCfeatures::uniform_pdf_transformed,
  n_particles = 10000,
  mcmc_trials = 10,
  discrepancy_final = 0,
  a_disc = 0.6,
  a_like = 0.3,
  c = 0.01,
  p_acc_min = 1e-04,
  n_cores = 1L
)
}
\arguments{
\item{upper}{a vector of upper parameter bounds. If using the Gaussian log-likelihood function (\link[SMCfeatures]{log_likelihood_Gaussian}), the last number of this array describes the standard deviation.}

\item{lower}{a vector of lower parameter bounds. If using the Gaussian log-likelihood function (\link[SMCfeatures]{log_likelihood_Gaussian}), the last number of this array describes the standard deviation.}

\item{simulate_model}{a function that simulates the model outputs needed to assess if constraints are met}

\item{calculate_discrepancy}{a function that calculates discrepancy score. See  \link[SMCfeatures]{discrepancy_logistic_growth} for an example.}

\item{simulate_data}{a function that simulates the model outputs needed to assess likelihood of producing data}

\item{input_data}{a vector of input data. Default NA.}

\item{output_data}{a vector of output data. Default NA.}

\item{calculate_log_likelihood}{log likelihood function. Default \link[SMCfeatures]{log_likelihood_Gaussian}.}

\item{include_expert_constraints}{boolean indicating whether expert ellicited constraints should be included. Default TRUE}

\item{include_data_constraints}{boolean indicating whether data should be included. Default TRUE. If input_data or output_data is NA, then include_data_constraints is set to FALSE}

\item{trans_f}{transform of prior parameter space to ensure unbounded support for MCMC sampling. Default \link[SMCfeatures]{uniform_transform}}

\item{trans_finv}{inverse of trans_f function. Default \link[SMCfeatures]{uniform_transform_inverse}}

\item{n_particles}{Number of desired ensemble members. Default to 10000}

\item{mcmc_trials}{number of MCMC steps to try before selecting appropriate number. Default 10}

\item{discrepancy_final}{target discrepancy threshold. Default 0. If zero, p_acc_min is used to determine stopping criteria.}

\item{a_disc}{tuning parameter for adaptive selection of discrepancy threshold sequence. Default 0.6}

\item{a_like}{tuning parameter for adaptive selection of likelihood ESS sequence. Default 0.3}

\item{c}{tuning parameter for choosing the number of MCMC iterations in move step. Default 0.01}

\item{p_acc_min}{minimum acceptable acceptance rate in the MCMC interations before exit. Default 0.0001}

\item{n_cores}{Number of cores desired to be used for sampling. Default set to 1 core (sequential sampling).}

\item{prior_sampler}{sampling function that generates random vectors from the joint prior distribution. Default \link[SMCfeatures]{uniform_sampler}.}

\item{prior_pdf}{joint probability density function. Default \link[SMCfeatures]{uniform_pdf_transformed}}
}
\value{
vector of transformed parameters
}
\description{
Main function for any parameterisation of expert elicited constraints
}
