% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMCfeatures.R
\name{SMCfeatures}
\alias{SMCfeatures}
\title{SMCfeatures}
\usage{
SMCfeatures(
  define_args = SMCfeatures::define_args_logistic_growth,
  include_expert_constraints = TRUE,
  include_data_constraints = TRUE,
  n_particles = 10000,
  mcmc_trials = 10,
  discrepancy_final = 0,
  a_disc = 0.6,
  a_like = 0.3,
  c = 0.01,
  p_acc_min = 1e-04,
  n_cores = 1L
)
}
\arguments{
\item{define_args}{function describing the arguments of the case study. Defaul \link[SMCfeatures]{define_args_logistic_growth}}

\item{include_expert_constraints}{boolean indicating whether expert ellicited constraints should be included. Default TRUE}

\item{include_data_constraints}{boolean indicating whether data should be included. Default TRUE}

\item{n_particles}{Number of desired ensemble members. Default to 10000}

\item{mcmc_trials}{number of MCMC steps to try before selecting appropriate number. Default 10}

\item{discrepancy_final}{target discrepancy threshold. Default 0. If zero, p_acc_min is used to determine stopping criteria.}

\item{a_disc}{tuning parameter for adaptive selection of discrepancy threshold sequence.}

\item{a_like}{tuning parameter for adaptive selection of likelihood ESS sequence}

\item{c}{tuning parameter for choosing the number of MCMC iterations in move step. Default 0.01}

\item{p_acc_min}{minimum acceptable acceptance rate in the MCMC interations before exit. Default 0.0001}

\item{n_cores}{Number of cores desired to be used for sampling. Default set to 1 core (sequential sampling).}
}
\value{
vector of transformed parameters
}
\description{
Main function for any parameterisation of expert elicited constraints
}
