% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_pdf_transformed.R
\name{uniform_pdf_transformed}
\alias{uniform_pdf_transformed}
\title{density of transformed uniform distribution}
\usage{
uniform_pdf_transformed(theta_trans)
}
\arguments{
\item{theta_trans}{vector of parameters to transform}
}
\value{
vector of transformed parameters
}
\description{
Density of transformed uniform distribution as defined in \link[SMCfeatures]{uniform_transform}
}
